requirements.txt
textual>=0.35.0
rich>=13.0.0
pyyaml>=6.0
click>=8.1
pygments>=2.15
pytest>=7.4
pytest-asyncio>=0.21.0
